+++ bowling/CMakeLists.txt
@@ -57,7 +57,7 @@ set(SOURCES
     # ${SOURCE_DIR}/argumentParser.cpp
     # ${SOURCE_DIR}/frameParser.cpp
     # ${SOURCE_DIR}/tokenizer.cpp
-    ${SOURCE_DIR}/Printer.cpp
+    ${SOURCE_DIR}/printer.cpp
     ${SOURCE_DIR}/consoleStream.cpp
     ${SOURCE_DIR}/fileStream.cpp
