+++ bowling/inc/printer.h
@@ -3,9 +3,6 @@
 #include "printableData.hpp"
 #include "printerStream.hpp"
 
-#include <ostream>
-#include <sstream>
-#include <fstream>
 #include <string>
 #include <vector>
 
@@ -17,8 +14,8 @@ public:
 
 private:
     void printHeader(LaneStruct & lane);
-    void print(std::string text);
-    void print(size_t value);
+    void printText(std::string text);
+    void printText(size_t value);
     std::string parseStatus(const Status status);
     void printPlayers(LaneStruct & lane);
     void printPlayer(Player & player);
